requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
PyMuPDF==1.23.26
numpy==1.24.3
scipy==1.10.1
nltk==3.8.1
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38
langchain-google-genai==1.0.3
google-generativeai>=0.5.2,<0.6.0
faiss-cpu==1.7.4
python-multipart==0.0.6
extract-msg==0.45.0

